requirement.txt
falsk
flask-sqlalchemy
flask-login
flask-bcrypt
flask-wtf
email_validator